libmemdraw: apply erik quanstros fix for sign preserving in byteaddr()

+++ sys/src/libmemdraw/alloc.c
@@ -143,8 +143,7 @@ byteaddr(Memimage *i, Point p)
 {
 	uchar *a;
 
-	a = i->data->bdata+i->zero+sizeof(ulong)*p.y*i->width;
-
+	a = i->data->bdata+i->zero+(int)(sizeof(ulong)*p.y*i->width);
 	if(i->depth < 8){
 		/*
 		 * We need to always round down,
